chore(ui): rename MewTimestamp to BaseTimestamp
